+++ server/src/server.ts
@@ -105,6 +105,9 @@ function findStartupFiles(workspaces: WorkspaceFolder[]): void {
 }
 
 function indexProject(pathsToProjectFiles: string[]): void {
+	if (pathsToProjectFiles.length == 0) {
+		projectIndex.updateProjectIsIndexed(true); // No startup.txt file to index
+	}
 	pathsToProjectFiles.map(async (path) => {
 		// TODO handle multiple startup.txt files in multiple directories
 
@@ -127,6 +130,8 @@ function indexProject(pathsToProjectFiles: string[]): void {
 		const promises = scenePaths.map(x => indexFile(x));
 
 		await Promise.all(promises);
+
+		projectIndex.updateProjectIsIndexed(true);
 
 		// Revalidate all open text documents
 		documents.all().forEach(doc => validateTextDocument(doc, projectIndex));